caelus_backend/requirements.txt
Flask
pytesseract
pdf2image
Pillow
python-dotenv
gunicorn